--- frontend/src/components/UserSettings.jsx
@@ -1,9 +1,0 @@
-import React from 'react'
-
-const UserSettings = () => {
-  return (
-    <div>UserSettings</div>
-  )
-}
-
-export default UserSettings
